moving core electra validator functions to helper to remove circular dependencies in other PRs
