+++ beacon-chain/core/electra/BUILD.bazel
@@ -2,10 +2,7 @@ load("@prysm//tools/go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "upgrade.go",
-        "validator.go",
-    ],
+    srcs = ["upgrade.go"],
     importpath = "github.com/prysmaticlabs/prysm/v5/beacon-chain/core/electra",
     visibility = ["//visibility:public"],
     deps = [
@@ -26,10 +23,7 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = [
-        "upgrade_test.go",
-        "validator_test.go",
-    ],
+    srcs = ["upgrade_test.go"],
     deps = [
         ":go_default_library",
         "//beacon-chain/core/helpers:go_default_library",
